refactor: rename and split app data and app action files
