+++ src/routines/index.ts
@@ -1,2 +1,3 @@
-export * from './apps';
+export * from './appData';
 export * from './appSetting';
+export * from './appAction';
