+++ src/routines/postMessage.ts
@@ -1,0 +1,4 @@
+import createRoutine from './utils';
+
+export const getAuthTokenRoutine = createRoutine('GET_AUTH_TOKEN');
+export const getLocalContextRoutine = createRoutine('GET_LOCAL_CONTEXT');
